requirements.txt
fastapi[standard]==0.104.1
uvicorn[standard]==0.24.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.30.0
psycopg2-binary==2.9.9
sentry-sdk[fastapi]==2.42.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.1.1
scikit-learn==1.5.2
joblib==1.4.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0

